% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat1.R
\name{cat1}
\alias{cat1}
\title{One categorical variable}
\usage{
cat1(y, l = "Overall", u = TRUE, ulab = "&nbsp;Missing", nam = "")
}
\arguments{
\item{y}{categorical variable to summarize}

\item{l}{Column label. Defaults to OVerall}

\item{u}{Include row for missing values? Defaults to TRUE}

\item{ulab}{Label for missing value row. Defaults to Missing}

\item{nam}{Variable label. Defaults to ""}
}
\description{
This function allows you to calculate variable summaries for a categorical variable for tables.
}
\examples{
cat1()
}
