% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat2.R
\name{cat2}
\alias{cat2}
\title{Two Categorical Variables}
\usage{
cat2(x, y, u = TRUE, ulab = "Missing", r = 2, nam = "")
}
\arguments{
\item{x}{Grouping variable}

\item{y}{Variable to summarize per group}

\item{u}{Include row for missing values? Defaults to TRUE}

\item{ulab}{Label for missing values}

\item{r}{Values, either 1 or 2 default - columns sum to 100 \%}

\item{nam}{Label of variable summarized. Defaults to ""}
}
\description{
This function allows you to summarize a categorical variable by a second grouping variable for tables.
}
\examples{
cat2()
}
\keyword{table1}
