% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/con1.R
\name{con1}
\alias{con1}
\title{One Continuous Variable}
\usage{
con1(y, l = "Overall", prmr = FALSE, nam = "")
}
\arguments{
\item{y}{Variable to summarize}

\item{l}{Label for summary column. Defaults to "Overall"}

\item{prmr}{Include median and range? Defaults to FALSE}

\item{nam}{Label of summary variable}
}
\description{
This function allows you to summarize a continuous variable for tables
}
\examples{
con1()
}
\keyword{table1}
