% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/con2.R
\name{con2}
\alias{con2}
\title{Grouped Continuous Variable}
\usage{
con2(x, y, d = 1, nam = "", prmr = TRUE)
}
\arguments{
\item{x}{Grouping variable}

\item{y}{Variablet o summarize per group}

\item{d}{Digits to print. Defaults to 1}

\item{nam}{Lable of characteristic column. Defaults to ""}

\item{prmr}{Include median and range? Defaults to FALSE.}
}
\description{
This function allows you to summarize a continuous variable by group for tables
}
\examples{
con2()
}
