% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_logistic_data.R
\name{make_logistic_data}
\alias{make_logistic_data}
\title{A table function for logistic regression models}
\usage{
make_logistic_data(m, d = 4, alpha = 0.95)
}
\arguments{
\item{m}{logistic model object}

\item{d}{digits to round table. Defaults to 4}

\item{alpha}{confidence level. Defaults to 0.95}
}
\description{
This function allows you to make nice summary tables for logistic regression models.
}
\examples{
make_logistic_data()
}
