% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_UMN.R
\name{theme_UMN}
\alias{theme_UMN}
\title{A ggplot2 theme for UMN publication use.}
\usage{
theme_UMN(base_size = 20, base_family = "serif")
}
\arguments{
\item{base_size}{Font size}

\item{base_family}{Font}
}
\description{
This theme makes a nice publishable figure with ggplot2.
}
\keyword{ggplot2}
\keyword{theme}
