% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{UMN_colors}
\alias{UMN_colors}
\title{UMN Brand standard color look up}
\format{
\subsection{\code{UMN_colors}}{

A data frame with 12 rows and 3 columns:
\describe{
\item{Name}{Color name}
\item{Type}{Primary or Secondary color}
\item{HEX}{HEX color code}
...
}
}
}
\source{
\url{https://university-relations.umn.edu/resources/colors-and-type}
}
\usage{
UMN_colors
}
\description{
A subset of data from the World Health Organization Global Tuberculosis
Report ...
}
\keyword{datasets}
